+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 - ♿️(frontend) add sr-only format to export download button #2088
 - ♿️(frontend) announce formatting shortcuts for screen readers #2070
 - ✨(frontend) add markdown copy icon for Copy as Markdown option #2096
+- 💄(frontend) improve comments highlights
 
 ### Fixed